src/main/java/code/problems/linkedlists/ReverseALinkedList.java
package code.problems.linkedlists;

import java.util.List;

/*
Problem:
Given the head of a singly linked list, reverse the list, and return the reversed list.

Constraints:
The number of nodes in the list is the range [0, 5000].
-5000 <= Node.val <= 5000
 */
public class ReverseALinkedList {

    public static ListNode reverse(ListNode head){
        ListNode prev = null;
        ListNode curr = head;

        while (curr != null){
            ListNode next = curr.next;

            curr.next = prev;
            prev = curr;
            curr = next;
        }

        return prev;
    }
}
